requirements.txt
certifi==2018.8.24
chardet==3.0.4
Click==7.0
cloudpickle==0.5.5
cycler==0.10.0
dask==0.18.2
decorator==4.3.0
docutils==0.14
Flask==1.0.2
graphviz==0.9
httpie==0.9.9
idna==2.7
imageio==2.4.1
itsdangerous==0.24
Jinja2==2.10
kiwisolver==1.0.1
lockfile==0.12.2
luigi==2.7.6
MarkupSafe==1.0
matplotlib==2.2.3
networkx==2.1
numpy==1.15.1
pandas==0.23.4
Pillow==5.2.0
Pygments==2.2.0
PyOpenGL==3.1.0
pyparsing==2.2.0
python-daemon==2.2.0
python-dateutil==2.7.3
pytz==2018.5
PyWavelets==0.5.2
requests==2.20.0
scikit-image==0.14.0
scikit-learn==0.19.2
scipy==1.1.0
seaborn==0.9.0
six==1.11.0
sklearn==0.0
toolz==0.9.0
torch==0.4.1
torchvision==0.2.1
tornado==4.5.3
urllib3==1.23
visvis==1.11.1
waitress==1.1.0
Werkzeug==0.14.1
